infrastructure/requirements.txt
# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
# SPDX-License-Identifier: MIT-0

aws-cdk-lib==2.199.0
constructs>=10.0.0,<11.0.0
black
boto3
s3fs
pytest==6.2.5
python-dotenv